Rename interactions
