+++ api/hasura/event_triggers/eventManager.ts
@@ -1,6 +1,6 @@
 import { VercelApiHandler, VercelRequest, VercelResponse } from '@vercel/node';
 
-import activityHandler from '../../../api-lib/event_triggers/activity/activityHandler';
+import activityHandler from '../../../api-lib/event_triggers/activity/index';
 import checkNomineeDiscord from '../../../api-lib/event_triggers/checkNomineeDiscord';
 import checkNomineeDiscordBot from '../../../api-lib/event_triggers/checkNomineeDiscordBot';
 import checkNomineeTelegram from '../../../api-lib/event_triggers/checkNomineeTelegram';
@@ -31,9 +31,9 @@ import { verifyHasuraRequestMiddleware } from '../../../api-lib/validate';
 type HandlerDict = { [handlerName: string]: VercelApiHandler };
 
 const HANDLERS: HandlerDict = {
-  activityHandlerUserInsert: activityHandler,
-  activityHandlerEpochInsert: activityHandler,
-  activityHandlerContributionInsert: activityHandler,
+  activityUserInsert: activityHandler,
+  activityEpochInsert: activityHandler,
+  activityContributionInsert: activityHandler,
   checkNomineeDiscord,
   checkNomineeDiscordBot,
   checkNomineeTelegram,
