+++ hasura/metadata/databases/default/tables/public_contributions.yaml
@@ -90,7 +90,7 @@ select_permissions:
               _is_null: true
       allow_aggregations: true
 event_triggers:
-  - name: activityHandlerContributionInsert
+  - name: activityContributionInsert
     definition:
       enable_manual: false
       insert:
